# Conflicts:
#	target/classes/com/bcopstein/adaptadores/controllers/SvFachadaRemota.class
#	target/classes/com/bcopstein/adaptadores/repositories/IProdutoRepositoryJPA.class
#	target/classes/com/bcopstein/adaptadores/repositories/ProdutoRepository.class
#	target/classes/com/bcopstein/negocio/entities/Produto.class
#	target/classes/com/bcopstein/negocio/repositories/IProdutoRepository.class

+++ src/main/java/com/bcopstein/adaptadores/controllers/SvFachadaRemota.java
@@ -32,7 +32,7 @@ public class SvFachadaRemota {
 
   @GetMapping("/autorizacao")
   @CrossOrigin(origins = "*")
-  public boolean podeVender(@RequestParam final Integer codProd,
+  public boolean podeVender(@RequestParam final Long codProd,
                             @RequestParam final Integer qtdade) {
     return service.podeVender(codProd, qtdade);
   }
